heartTaskCode/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
anyio=4.4.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
attrs=23.2.0=pypi_0
babel=2.15.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
bleach=6.1.0=pypi_0
brotli-python=1.0.9=py312hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.7.2=haa95532_0
certifi=2024.7.4=py312haa95532_0
cffi=1.16.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
clip=0.2.0=pypi_0
colorama=0.4.6=py312haa95532_0
comm=0.2.2=pypi_0
contourpy=1.2.1=pypi_0
cuda-cccl=12.4.127=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.127=0
cuda-opencl-dev=12.4.127=0
cuda-profiler-api=12.4.127=0
cuda-runtime=12.1.0=0
cycler=0.12.1=pypi_0
cython=3.0.10=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pypi_0
docker-pycreds=0.4.0=pypi_0
et-xmlfile=1.1.0=pypi_0
executing=2.0.1=pypi_0
expat=2.6.2=hd77b12b_0
ez-setup=0.9=pypi_0
fastjsonschema=2.19.1=pypi_0
filelock=3.13.1=py312haa95532_0
fonttools=4.51.0=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=ha860e81_0
frozenlist=1.4.1=pypi_0
fsspec=2024.5.0=pypi_0
ftfy=6.2.0=pypi_0
future=1.0.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
grpcio=1.63.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.23.2=pypi_0
idna=3.7=py312haa95532_0
imageio=2.35.1=pypi_0
install=1.3.5=pypi_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.29.4=pypi_0
ipython=8.24.0=pypi_0
ipywidgets=8.1.2=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=py312haa95532_0
jinja2=3.1.3=py312haa95532_0
joblib=1.4.2=pypi_0
jpeg=9e=h2bbff1b_1
json5=0.9.25=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.6.2=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyter_client=8.6.0=py312haa95532_0
jupyter_core=5.7.2=py312haa95532_0
jupyterlab=4.2.1=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.2=pypi_0
jupyterlab-widgets=3.0.10=pypi_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.4=pypi_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.5.147=0
libcurand-dev=10.3.5.147=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_1
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libpng=1.6.39=h8cc25b3_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp-base=1.3.2=h2bbff1b_0
lightning-bolts=0.4.0=pypi_0
lightning-utilities=0.11.2=pypi_0
lz4-c=1.9.4=h2bbff1b_1
madgrad=1.1=pypi_0
markdown=3.6=pypi_0
markupsafe=2.1.3=py312h2bbff1b_0
matplotlib=3.9.0=pypi_0
matplotlib-inline=0.1.7=pypi_0
mistune=3.0.2=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py312h2bbff1b_1
mkl_fft=1.3.8=py312h2bbff1b_0
mkl_random=1.2.4=py312h59b6b97_0
mpmath=1.3.0=py312haa95532_0
multidict=6.0.5=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
nest-asyncio=1.6.0=py312haa95532_0
networkx=3.1=py312haa95532_0
nibabel=5.2.1=pypi_0
notebook=7.2.0=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.26.4=py312hfd52020_0
numpy-base=1.26.4=py312h4dde369_0
open-clip-torch=2.20.0=pypi_0
opencv-python=4.10.0.84=pypi_0
opencv-python-headless=4.8.0.76=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openpyxl=3.1.5=pypi_0
openssl=3.0.14=h827c3e9_0
overrides=7.7.0=pypi_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pillow=10.0.0=pypi_0
pip=24.0=py312haa95532_0
platformdirs=4.2.2=pypi_0
plotly=5.24.1=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.43=py312haa95532_0
prompt_toolkit=3.0.43=hd3eb1b0_0
protobuf=3.20.3=pypi_0
psutil=5.9.8=pypi_0
pure_eval=0.2.2=pyhd3eb1b0_0
py-cpuinfo=9.0.0=pypi_0
pycparser=2.22=pypi_0
pydeprecate=0.3.1=pypi_0
pydicom=2.4.4=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py312haa95532_0
python=3.12.3=h1d929f7_1
python-dateutil=2.9.0post0=py312haa95532_2
python-json-logger=2.0.7=pypi_0
pytorch=2.2.2=py3.12_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-lightning=1.5.5=pypi_0
pytorch-metric-learning=2.6.0=pypi_0
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pywin32=306=pypi_0
pywinpty=2.0.13=pypi_0
pyyaml=6.0.1=py312h2bbff1b_0
pyzmq=26.0.3=pypi_0
qtconsole=5.5.2=pypi_0
qtpy=2.4.1=pypi_0
referencing=0.35.1=pypi_0
regex=2024.5.15=pypi_0
requests=2.31.0=py312haa95532_1
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.18.1=pypi_0
safetensors=0.4.3=pypi_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.4.2=pypi_0
scipy=1.12.0=pypi_0
seaborn=0.13.2=pypi_0
send2trash=1.8.3=pypi_0
sentencepiece=0.2.0=pypi_0
sentry-sdk=2.2.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=69.5.1=pypi_0
simpleitk=2.3.1=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.45.3=h2bbff1b_0
stack-data=0.6.3=pypi_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.12=py312haa95532_0
tbb=2021.8.0=h59b6b97_0
tenacity=9.0.0=pypi_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.9.20=pypi_0
timm=0.9.5=pypi_0
tinycss2=1.3.0=pypi_0
tk=8.6.14=h0416ee5_0
tokenizers=0.19.1=pypi_0
torchaudio=2.2.2=pypi_0
torchmetrics=1.5.1=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.17.2=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.5=pypi_0
traitlets=5.14.3=py312haa95532_0
transformers=4.41.2=pypi_0
types-python-dateutil=2.9.0.20240316=pypi_0
typing_extensions=4.11.0=py312haa95532_0
tzdata=2024.1=pypi_0
ultralytics=8.3.15=pypi_0
ultralytics-thop=2.0.9=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.1=py312haa95532_0
vc=14.2=h2eaa2aa_1
vs2015_runtime=14.29.30133=h43f2093_3
wandb=0.17.0=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py312haa95532_0
widgetsnbextension=4.0.10=pypi_0
win_inet_pton=1.1.0=py312haa95532_0
xlrd=2.0.1=pypi_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
yarl=1.9.4=pypi_0
zeromq=4.3.5=hd77b12b_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.5=hd43e919_2
